로그인, 로그아웃, 회원가입, 주소를 위도,경도변환 완료 토큰 재발급, 블랙리스트 완료 이메일 인증, 닉네임 체크 완료

+++ back/cors/src/main/java/ml/market/cors/domain/member/entity/MemberDAO.java
@@ -37,10 +37,10 @@ public class MemberDAO implements UserDetails{
     private String address;
 
     @Column(name = "latitude")
-    private int latitude;
+    private double latitude;
 
     @Column(name = "longitude")
-    private int longitude;
+    private double longitude;
 
     @Column(name = "nickname")
     private String nickname;
@@ -60,7 +60,7 @@ public class MemberDAO implements UserDetails{
         this.member_id = member_id;
     }
 
-    public MemberDAO(String email, MemberRole role, String password, String address, int latitude, int longitude, String nickname
+    public MemberDAO(String email, MemberRole role, String password, String address, double latitude, double longitude, String nickname
     ,eSubscriberType subscriberType) {
         this.email = email;
         this.role = role;
